+++ CHANGELOG.md
@@ -188,3 +188,4 @@
     - Bug-088424 - Sorting filters are not working properly on the Banners listing page on the admin end
     - Bug-088419 - Item counts on the collections listing page are not correct
     - Bug-088422 - Banners added by the admin are not visible on the product details page on the admin end
+    - Bug-088545 - Multiple banners list appears in the banner collection 1